refactor(phase2): split authorization.rs monolith into 6 focused sub-modules
The 4,448-line `rest_macro_core/src/authorization.rs` is replaced by a
proper module directory `src/authorization/` with one responsibility per file:

  types.rs      (1045 L)  — all data types, impl blocks, ResolvedRuntimeAssignments,
                             AuthorizationModel, AuthorizationRuntime, ResourceAuthorization
  engine.rs     (1278 L)  — condition evaluation, exists trace helpers,
                             hybrid enforcement simulation, assignment tracing,
                             parse_runtime_assignment_timestamp
  migrations.rs   (92 L)  — authorization_runtime_migration_sql DDL
  db_ops.rs      (560 L)  — CRUD for runtime assignments + events, row parsing,
                             error detection helpers (is_missing/outdated table)
  handlers.rs    (175 L)  — 7 actix-web endpoint handlers (pub(super))
  routing.rs      (60 L)  — authorization_management_routes[_at]
  mod.rs        (1382 L)  — sub-module declarations, full pub-use re-exports, test suite

Public API is 100% preserved via re-exports in mod.rs.
`cargo test -p rest_macro_core`: 56 tests ✓, 0 failures.
