+++ crates/rest_macro_core/src/authorization/migrations.rs
@@ -1,0 +1,92 @@
+use crate::auth::AuthDbBackend;
+
+use super::types::{
+    AUTHORIZATION_RUNTIME_ASSIGNMENT_EVENT_TABLE, AUTHORIZATION_RUNTIME_ASSIGNMENT_TABLE,
+};
+
+pub fn authorization_runtime_migration_sql(backend: AuthDbBackend) -> String {
+    let id_column = match backend {
+        AuthDbBackend::Sqlite | AuthDbBackend::Postgres => "id TEXT PRIMARY KEY",
+        AuthDbBackend::Mysql => "id VARCHAR(191) PRIMARY KEY",
+    };
+    let target_kind_column = match backend {
+        AuthDbBackend::Sqlite | AuthDbBackend::Postgres => "target_kind TEXT NOT NULL",
+        AuthDbBackend::Mysql => "target_kind VARCHAR(32) NOT NULL",
+    };
+    let target_name_column = match backend {
+        AuthDbBackend::Sqlite | AuthDbBackend::Postgres => "target_name TEXT NOT NULL",
+        AuthDbBackend::Mysql => "target_name VARCHAR(191) NOT NULL",
+    };
+    let scope_name_column = match backend {
+        AuthDbBackend::Sqlite | AuthDbBackend::Postgres => "scope_name TEXT NOT NULL",
+        AuthDbBackend::Mysql => "scope_name VARCHAR(191) NOT NULL",
+    };
+    let scope_value_column = match backend {
+        AuthDbBackend::Sqlite | AuthDbBackend::Postgres => "scope_value TEXT NOT NULL",
+        AuthDbBackend::Mysql => "scope_value VARCHAR(191) NOT NULL",
+    };
+    let user_id_column = match backend {
+        AuthDbBackend::Sqlite => "user_id INTEGER NOT NULL",
+        AuthDbBackend::Postgres | AuthDbBackend::Mysql => "user_id BIGINT NOT NULL",
+    };
+    let created_by_user_id_column = match backend {
+        AuthDbBackend::Sqlite => "created_by_user_id INTEGER",
+        AuthDbBackend::Postgres | AuthDbBackend::Mysql => "created_by_user_id BIGINT",
+    };
+    let created_at_column = match backend {
+        AuthDbBackend::Sqlite | AuthDbBackend::Postgres => "created_at TEXT NOT NULL",
+        AuthDbBackend::Mysql => "created_at VARCHAR(64) NOT NULL",
+    };
+    let expires_at_column = match backend {
+        AuthDbBackend::Sqlite | AuthDbBackend::Postgres => "expires_at TEXT",
+        AuthDbBackend::Mysql => "expires_at VARCHAR(64)",
+    };
+    let event_kind_column = match backend {
+        AuthDbBackend::Sqlite | AuthDbBackend::Postgres => "event_kind TEXT NOT NULL",
+        AuthDbBackend::Mysql => "event_kind VARCHAR(32) NOT NULL",
+    };
+    let assignment_id_column = match backend {
+        AuthDbBackend::Sqlite | AuthDbBackend::Postgres => "assignment_id TEXT NOT NULL",
+        AuthDbBackend::Mysql => "assignment_id VARCHAR(191) NOT NULL",
+    };
+    let reason_column = match backend {
+        AuthDbBackend::Sqlite | AuthDbBackend::Postgres => "reason TEXT",
+        AuthDbBackend::Mysql => "reason TEXT",
+    };
+
+    format!(
+        "-- Generated by very_simple_rest for runtime authorization assignments.\n\n\
+         CREATE TABLE {table} (\n\
+             {id_column},\n\
+             {user_id_column},\n\
+             {created_by_user_id_column},\n\
+             {created_at_column},\n\
+             {expires_at_column},\n\
+             {target_kind_column},\n\
+             {target_name_column},\n\
+             {scope_name_column},\n\
+             {scope_value_column}\n\
+         );\n\n\
+         CREATE INDEX idx_{table}_user ON {table} (user_id);\n\
+         CREATE INDEX idx_{table}_scope ON {table} (scope_name, scope_value);\n\n\
+         CREATE TABLE {event_table} (\n\
+             {id_column},\n\
+             {assignment_id_column},\n\
+             {user_id_column},\n\
+             {created_by_user_id_column},\n\
+             {created_at_column},\n\
+             {event_kind_column},\n\
+             {target_kind_column},\n\
+             {target_name_column},\n\
+             {scope_name_column},\n\
+             {scope_value_column},\n\
+             {expires_at_column},\n\
+             {reason_column}\n\
+         );\n\n\
+         CREATE INDEX idx_{event_table}_user ON {event_table} (user_id, created_at);\n\
+         CREATE INDEX idx_{event_table}_assignment ON {event_table} (assignment_id, created_at);\n",
+        table = AUTHORIZATION_RUNTIME_ASSIGNMENT_TABLE,
+        event_table = AUTHORIZATION_RUNTIME_ASSIGNMENT_EVENT_TABLE,
+        assignment_id_column = assignment_id_column,
+    )
+}
